+++ src/azure-pipelines-vscode-ext/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Support native folder path in settings / prompt
 - Allow to show a folder picker if leaving the path empty
 - Ask to save the changed repositories
+- Disable errors for autocomplete, semantic tokens and hover
 
 ### v0.2.6 (Preview)
 - Includes experimental bug fixes for dynamic each expressions https://github.com/ChristopherHX/runner.server/issues/278 and https://github.com/ChristopherHX/runner.server/issues/514